refactor: aligns blueprint implementations, reports, and setup update orchestration.
