+++ src/core/blueprint/blueprint-implementation.ts
@@ -1,0 +1,3 @@
+export interface BlueprintImplementation<TOptions, TResult> {
+  apply(options: TOptions): Promise<TResult>;
+}
