+++ src/core/blueprint/blueprint-report.ts
@@ -1,0 +1,7 @@
+import type { BaseContext } from 'clipanion';
+
+export type OutputWriter = Pick<BaseContext['stdout'], 'write'>;
+
+export interface BlueprintReport {
+  writeOutput(output: OutputWriter): void;
+}
