+++ src/core/interactive-merge/index.ts
@@ -3,9 +3,3 @@ export {
   UserAbortError,
   type InteractiveDiffSelectorOptions,
 } from './update-interactive';
-export {
-  runSetupUpdate,
-  writeSetupUpdateSummary,
-  type SetupUpdateOptions,
-  type SetupUpdateOutput,
-} from './merge-orchestrator';
